+++ source/deprecated/omni.isaac.manipulators/premake5.lua
@@ -1,0 +1,14 @@
+local ext = get_current_extension_info()
+ext.target_dir = deprecated_exts_path.."/"..ext.id
+
+
+project_ext(ext)
+
+
+repo_build.prebuild_link {
+    { "docs", ext.target_dir.."/docs" },
+    { "data", ext.target_dir.."/data" },
+    { "python/controllers", ext.target_dir.."/omni/isaac/manipulators/controllers" },
+    { "python/manipulators", ext.target_dir.."/omni/isaac/manipulators/manipulators" },
+    { "python/grippers", ext.target_dir.."/omni/isaac/manipulators/grippers" },
+}