+++ source/extensions/isaacsim.core.api/python/scripts/tasks/stacking.py
@@ -173,7 +173,7 @@ class Stacking(ABC, BaseTask):
 
     def post_reset(self) -> None:
         """[summary]"""
-        from omni.isaac.manipulators.grippers.parallel_gripper import ParallelGripper
+        from isaacsim.robot.manipulators.grippers.parallel_gripper import ParallelGripper
 
         if isinstance(self._robot.gripper, ParallelGripper):
             self._robot.gripper.set_joint_positions(self._robot.gripper.joint_opened_positions)