+++ source/standalone_examples/testing/omni.isaac.franka/torque_control.py
@@ -16,7 +16,7 @@ from isaacsim.core.api import World
 from isaacsim.core.api.controllers.base_controller import BaseController
 from isaacsim.core.api.tasks import BaseTask
 from isaacsim.core.api.utils.types import ArticulationAction
-from omni.isaac.franka import Franka
+from isaacsim.robot.manipulators.examples.franka import Franka
 
 my_world = World(stage_units_in_meters=1.0)